Add Author entity & admin UI; use Long quantities
feat: Bổ sung quản lý Author, liên kết Product-Author và chuẩn hóa kiểu Long cho quantity/stock

Thêm Author model, AuthorRepository, AuthorController.
Thêm trang admin-authors.html với CRUD và tìm kiếm.
Cập nhật Product liên kết Author qua authorEntity, vẫn giữ author String cũ để tương thích ngược.
Bổ sung getDisplayAuthor() để hiển thị tác giả phù hợp.
Cập nhật ProductController:
Nhận authorId khi lưu.
Tự động active sản phẩm mới.
Trả JSON khi lưu sản phẩm.
Điều chỉnh logic xóa/bật-tắt trạng thái.
Lưu gallery images.
Đổi kiểu quantity/stock sang Long trong CartDetail, OrderDetail, Product.
Cập nhật Cart/Checkout/Favourite/Index/Order controller để dùng long-stream và phép toán Long.
Mở rộng AuthorRepository:
Thêm truy vấn, gồm đếm số sản phẩm theo tác giả.
Hỗ trợ soft-delete/restore tác giả khi còn liên kết sản phẩm.
Lưu ý:
Cần migration DB/schema: tạo bảng authors, thêm FK author_id, và có thể đổi kiểu quantity/stock.
Cần cập nhật frontend để xử lý response JSON mới của API lưu sản phẩm.

+++ src/main/java/com/poly/graduation_project/controller/CartController.java
@@ -84,7 +84,7 @@ public class CartController {
                         ? java.math.BigDecimal.ZERO
                         : new java.math.BigDecimal("30000");
 
-        int totalQuantity = sortedItems.stream().mapToInt(CartDetail::getQuantity).sum();
+        Long totalQuantity = sortedItems.stream().mapToLong(CartDetail::getQuantity).sum();
 
         model.addAttribute("cartItems",      sortedItems);
         model.addAttribute("hasActiveItems", hasActiveItems);
@@ -141,7 +141,7 @@ public class CartController {
                 .orElse(null);
 
         if (cartDetail != null) {
-            int newQty = cartDetail.getQuantity() + quantity;
+            Long newQty = cartDetail.getQuantity() + quantity;
             newQty = Math.min(newQty, product.getStockQuantity());
             cartDetail.setQuantity(newQty);
         } else {
@@ -191,11 +191,11 @@ public class CartController {
             return ResponseEntity.ok(response);
         }
 
-        int stock = cartDetail.getProduct().getStockQuantity();
+        Long stock = cartDetail.getProduct().getStockQuantity();
         if (quantity < 1) quantity = 1;
-        if (quantity > stock) quantity = stock;
+        if (quantity > stock.intValue()) quantity = stock.intValue();
 
-        cartDetail.setQuantity(quantity);
+        cartDetail.setQuantity(Long.valueOf(quantity));
         cartDetailRepository.save(cartDetail);
 
         java.math.BigDecimal lineTotal = cartDetail.getProduct().getPrice()
